requirements.txt
Django==4.0.2
django-bootstrap-datepicker-plus==4.0.0
openpyxl==3.0.9